adding swagger description

+++ ITI.FinalProject.WebAPI/Controllers/BranchesController.cs
@@ -5,6 +5,7 @@ using Application.Interfaces.ApplicationServices;
 using Domain.Entities;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using Swashbuckle.AspNetCore.Annotations;
 
 namespace ITI.FinalProject.WebAPI.Controllers
 {
@@ -17,13 +18,32 @@ namespace ITI.FinalProject.WebAPI.Controllers
         {
             branchServ = _branchServ;
         }
+
+        [SwaggerOperation(
+        Summary = "This Endpoint returns a list of branches",
+            Description = ""
+        )]
+        [SwaggerResponse(404, "There weren't any branches in the database", Type = typeof(void))]
+        [SwaggerResponse(200, "Returns A list of branches", Type = typeof(List<BranchDisplayDTO>))]
+
         [HttpGet]
         public async Task<ActionResult> getAllBranches()
         {
             List<BranchDisplayDTO> branches = await branchServ.GetAllObjects();
+            if(branches.Count == 0)
+            {
+                return NotFound();
+            }
             return Ok(branches);
 
         }
+         [SwaggerOperation(
+       Summary = "This Endpoint returns the specified branch",
+           Description = ""
+       )]
+        [SwaggerResponse(404, "The id that was given doesn't exist in the db", Type = typeof(void))]
+        [SwaggerResponse(200, "Returns the specified branch", Type = typeof(BranchDisplayDTO))]
+
         [HttpGet("id")]
         public async Task<ActionResult> getById(int id)
         {
@@ -33,6 +53,14 @@ namespace ITI.FinalProject.WebAPI.Controllers
             return Ok(branch);
 
         }
+        [SwaggerOperation(
+      Summary = "This Endpoint inserts a city element in the db",
+          Description = ""
+      )]
+        [SwaggerResponse(400, "Something went wrong, please check your request", Type = typeof(void))]
+        [SwaggerResponse(201, "Confirms that the city was inserted successfully", Type = typeof(void))]
+        [SwaggerResponse(500, "Something went wrong, please try again later", Type = typeof(void))]
+
         [HttpPost]
         public async Task<ActionResult> addBranch(BranchInsertDTO branch)
         {
@@ -41,25 +69,46 @@ namespace ITI.FinalProject.WebAPI.Controllers
             var result =await branchServ.InsertObject(branch);
             if (result.Succeeded)
             {
-                return Ok(result);
+                return Created();
             }
-            return BadRequest(result.Message);
+            return StatusCode(StatusCodes.Status500InternalServerError,result.Message);
 
 
         }
+        [SwaggerOperation(
+      Summary = "This Endpoint inserts a city element in the db",
+          Description = ""
+      )]
+        [SwaggerResponse(404, "The id that was given doesn't exist in the db", Type = typeof(void))]
+        [SwaggerResponse(200, "Confirms that the city was deleted successfully", Type = typeof(void))]
+        [SwaggerResponse(500, "Something went wrong, please try again later", Type = typeof(void))]
+
         [HttpDelete]
         public async Task<ActionResult> deleteBranch(int id)
         {
-
+            BranchDisplayDTO? branch = await branchServ.GetObjectWithoutTracking(c => c.id == id);
+            if (branch == null)
+            {
+                return NotFound();
+            }
             var result =await branchServ.DeleteObject(id);
             if (result.Succeeded)
             {
                 return Ok(result);
             }
-            return BadRequest(result.Message);
+            return StatusCode(StatusCodes.Status500InternalServerError,result.Message);
 
 
         }
+        [SwaggerOperation(
+       Summary = "This Endpoint updates the specified city",
+           Description = ""
+       )]
+        [SwaggerResponse(404, "The id that was given doesn't exist in the db", Type = typeof(void))]
+        [SwaggerResponse(400, "The id that was given doesn't equal the id in the given city object", Type = typeof(void))]
+        [SwaggerResponse(500, "Something went wrong, please try again later", Type = typeof(void))]
+        [SwaggerResponse(200, "Confirms that the city was updated successfully", Type = typeof(void))]
+
         [HttpPut("id")]
         public async Task<ActionResult> updateBranch(int id,BranchUpdateDTO branch)
         {
@@ -68,12 +117,18 @@ namespace ITI.FinalProject.WebAPI.Controllers
                 return BadRequest();
             }
 
+            BranchDisplayDTO? branchDisplay = await branchServ.GetObjectWithoutTracking(c => c.id == id);
+            if (branch == null)
+            {
+                return NotFound();
+            }
+
             var result=await branchServ.UpdateObject(branch);
             if(result.Succeeded)
             {
                 return Ok(branch);
             }
-            return BadRequest(result.Message);
+            return StatusCode(StatusCodes.Status500InternalServerError, result.Message);
         }
 
     }
