+++ ITI.FinalProject.WebAPI/Controllers/CitiesController.cs
@@ -5,6 +5,7 @@ using Application.Interfaces.ApplicationServices;
 using Domain.Entities;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using Swashbuckle.AspNetCore.Annotations;
 
 namespace ITI.FinalProject.WebAPI.Controllers
 {
@@ -17,13 +18,32 @@ namespace ITI.FinalProject.WebAPI.Controllers
         {
             CityServ = _CityServ;
         }
+
+        [SwaggerOperation(
+        Summary = "This Endpoint returns a list of cities",
+            Description = ""
+        )]
+        [SwaggerResponse(404, "There weren't any cities in the database", Type = typeof(void))]
+        [SwaggerResponse(200, "Returns A list of cities", Type = typeof(List<CityDisplayDTO>))]
         [HttpGet]
         public async Task<ActionResult> getAllCities()
         {
             List<CityDisplayDTO> Cities = await CityServ.GetAllObjects();
+            if(Cities.Count==0)
+            {
+                return NotFound();
+            }
             return Ok(Cities);
 
         }
+
+
+        [SwaggerOperation(
+       Summary = "This Endpoint returns the specified city",
+           Description = ""
+       )]
+        [SwaggerResponse(404, "The id that was given doesn't exist in the db", Type = typeof(void))]
+        [SwaggerResponse(200, "Returns the specified city", Type = typeof(CityDisplayDTO))]
         [HttpGet("id")]
         public async Task<ActionResult> getById(int id)
         {
@@ -33,6 +53,14 @@ namespace ITI.FinalProject.WebAPI.Controllers
             return Ok(City);
 
         }
+        [SwaggerOperation(
+       Summary = "This Endpoint inserts a city element in the db",
+           Description = ""
+       )]
+        [SwaggerResponse(400, "Something went wrong, please check your request", Type = typeof(void))]
+        [SwaggerResponse(201, "Confirms that the city was inserted successfully", Type = typeof(void))]
+        [SwaggerResponse(500, "Something went wrong, please try again later", Type = typeof(void))]
+
         [HttpPost]
         public async Task <ActionResult> addCity(CityInsertDTO City)
         {
@@ -41,31 +69,59 @@ namespace ITI.FinalProject.WebAPI.Controllers
             var result =await CityServ.InsertObject(City);
             if (result.Succeeded)
             {
-                return Ok(result);
+                return Created();
             }
-            return BadRequest(result.Message);
+            return StatusCode(StatusCodes.Status500InternalServerError,result.Message);
 
 
         }
+
+        [SwaggerOperation(
+      Summary = "This Endpoint inserts a city element in the db",
+          Description = ""
+      )]
+        [SwaggerResponse(404, "The id that was given doesn't exist in the db", Type = typeof(void))]
+        [SwaggerResponse(200, "Confirms that the city was deleted successfully", Type = typeof(void))]
+        [SwaggerResponse(500, "Something went wrong, please try again later", Type = typeof(void))]
+
         [HttpDelete]
         public async Task<ActionResult> deleteCity(int id)
         {
+            CityDisplayDTO? city =await CityServ.GetObjectWithoutTracking(c => c.id == id);
+            if(city == null)
+            {
+                return NotFound();
+            }
 
             var result =await CityServ.DeleteObject(id);
             if (result.Succeeded)
             {
                 return Ok(result);
             }
-            return BadRequest(result.Message);
+            return StatusCode(StatusCodes.Status500InternalServerError);
 
 
         }
+        [SwaggerOperation(
+       Summary = "This Endpoint updates the specified city",
+           Description = ""
+       )]
+        [SwaggerResponse(404, "The id that was given doesn't exist in the db", Type = typeof(void))]
+        [SwaggerResponse(400, "The id that was given doesn't equal the id in the given city object", Type = typeof(void))]
+        [SwaggerResponse(500, "Something went wrong, please try again later", Type = typeof(void))]
+        [SwaggerResponse(200, "Confirms that the city was updated successfully", Type = typeof(void))]
+
         [HttpPut("id")]
         public async Task<ActionResult> updateCity(int id, CityUpdateDTO city)
         {
             if (city == null || id != city.id)
             {
-                return BadRequest();
+                return BadRequest("Id doesn't match the id in the object");
+            }
+            CityDisplayDTO? cityDisplay = await CityServ.GetObjectWithoutTracking(c => c.id == id);
+            if (cityDisplay == null)
+            {
+                return NotFound();
             }
 
             var result = await CityServ.UpdateObject(city);
@@ -73,7 +129,7 @@ namespace ITI.FinalProject.WebAPI.Controllers
             {
                 return Ok(city);
             }
-            return BadRequest(result.Message);
+            return StatusCode(StatusCodes.Status500InternalServerError);
         }
 
     }
